backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
google-genai
google-generativeai
psycopg2-binary
alembic
